% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legacy_sites.R
\name{add_legacy_sites}
\alias{add_legacy_sites}
\title{Add legacy sites}
\usage{
add_legacy_sites(legacy, bbox, cellsize, grtsdb = connect_db())
}
\arguments{
\item{legacy}{A matrix with coordinates of the legacy sites.
One column for every dimension.}

\item{bbox}{A two-column matrix.
The first column has the minimum, the second the maximum values.
Rows represent the spatial dimensions.}

\item{cellsize}{The size of each cell.
Either a single value or one value for each dimension.}

\item{grtsdb}{the grtsdb object}
}
\description{
Add legacy sites
}
\seealso{
Other legacy: 
\code{\link{drop_legacy_sites}()},
\code{\link{extract_legacy_sample}()}
}
\concept{legacy}
