% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_db.R
\name{compact_db}
\alias{compact_db}
\title{Compact the database by removing the lower levels and all indices}
\usage{
compact_db(grtsdb = connect_db())
}
\arguments{
\item{grtsdb}{the grtsdb object}
}
\description{
Compact the database by removing the lower levels and all indices
}
\seealso{
Other utility: 
\code{\link{create_index}()},
\code{\link{has_index}()},
\code{\link{is_grtsdb}()},
\code{\link{n_level}()},
\code{\link{which_level}()}
}
\concept{utility}
