% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{connect to a grtsdb database}
\usage{
connect_db(db = getOption("grtsdb", "grts.sqlite"))
}
\arguments{
\item{db}{the name of the database}
}
\description{
connect to a grtsdb database
}
