% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legacy_sites.R
\name{drop_legacy_sites}
\alias{drop_legacy_sites}
\title{Drop the table with legacy sites for a given level}
\usage{
drop_legacy_sites(level, grtsdb = connect_db())
}
\arguments{
\item{level}{the required level}

\item{grtsdb}{the grtsdb object}
}
\description{
Drop the table with legacy sites for a given level
}
\seealso{
Other legacy: 
\code{\link{add_legacy_sites}()},
\code{\link{extract_legacy_sample}()}
}
\concept{legacy}
