% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_index.R
\name{has_index}
\alias{has_index}
\title{has a table the required index}
\usage{
has_index(level, grtsdb = connect_db())
}
\arguments{
\item{level}{the required level}

\item{grtsdb}{the grtsdb object}
}
\description{
has a table the required index
}
