% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level}
\alias{level}
\title{Calculate the required level based on a bounding box and a cellsize}
\usage{
level(bbox, cellsize)
}
\arguments{
\item{bbox}{two-column matrix; the first column has the minimum, the second the maximum values; rows represent the spatial dimensions}

\item{cellsize}{the size of each cell. Either a single value or one value for each dimension.}
}
\value{
the required level to cover the bbox using a grid with cellsize
}
\description{
Calculate the required level based on a bounding box and a cellsize
}
