% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_level.R
\name{n_level}
\alias{n_level}
\title{Calculate the required level based on a bounding box and a cellsize}
\usage{
n_level(bbox, cellsize)
}
\arguments{
\item{bbox}{A two-column matrix.
The first column has the minimum, the second the maximum values.
Rows represent the spatial dimensions.}

\item{cellsize}{The size of each cell.
Either a single value or one value for each dimension.}
}
\value{
the required level to cover the bbox using a grid with cellsize
}
\description{
Calculate the required level based on a bounding box and a cellsize
}
