% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dbDisconnect}
\title{Objects exported from other packages}
\seealso{
Other base: 
\code{\link{add_level}()},
\code{\link{connect_db}()},
\code{\link{extract_sample}()}
}
\concept{base}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DBI}{\code{\link[DBI]{dbDisconnect}}}
}}

