% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_level.R
\name{which_level}
\alias{which_level}
\title{Return a vector of level number which are available in the database}
\usage{
which_level(grtsdb = connect_db())
}
\arguments{
\item{grtsdb}{the grtsdb object}
}
\description{
Return a vector of level number which are available in the database
}
\seealso{
Other utility: 
\code{\link{compact_db}()},
\code{\link{create_index}()},
\code{\link{has_index}()},
\code{\link{is_grtsdb}()},
\code{\link{n_level}()}
}
\concept{utility}
