% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_level.R
\name{add_level}
\alias{add_level}
\title{Add a level to the grtsdb}
\usage{
add_level(grtsdb = connect_db(), bbox, cellsize, verbose = TRUE, level)
}
\arguments{
\item{grtsdb}{the grtsdb object}

\item{bbox}{two-column matrix; the first column has the minimum, the second the maximum values; rows represent the spatial dimensions}

\item{cellsize}{the size of each cell. Either a single value or one value for each dimension.}

\item{verbose}{Display progress}

\item{level}{the required level}
}
\description{
Add a level to the grtsdb
}
