% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_level.R
\name{add_level}
\alias{add_level}
\title{Add a level to the grtsdb}
\usage{
add_level(bbox, cellsize, grtsdb = connect_db(), verbose = TRUE, level)
}
\arguments{
\item{bbox}{A two-column matrix.
The first column has the minimum, the second the maximum values.
Rows represent the spatial dimensions.}

\item{cellsize}{The size of each cell.
Either a single value or one value for each dimension.}

\item{grtsdb}{the grtsdb object}

\item{verbose}{Display progress}

\item{level}{the required level}
}
\description{
Add a level to the grtsdb
}
