% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{connect to or create an SQLite database}
\usage{
connect_db(db = getOption("grtsdb", "grts.sqlite"))
}
\arguments{
\item{db}{the name or path of the database}
}
\description{
connect to or create an SQLite database
}
\seealso{
Other base: 
\code{\link{add_level}()},
\code{\link{extract_sample}()},
\code{\link{reexports}}
}
\concept{base}
