% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index.R
\name{create_index}
\alias{create_index}
\title{Add the index to the table}
\usage{
create_index(level, grtsdb = connect_db(), legacy = FALSE, ...)
}
\arguments{
\item{level}{the required level}

\item{grtsdb}{the grtsdb object}

\item{legacy}{Use legacy sites.
Defaults to \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[=add_level]{add_level}}
  \describe{
    \item{\code{verbose}}{Display progress}
    \item{\code{bbox}}{A two-column matrix.
The first column has the minimum, the second the maximum values.
Rows represent the spatial dimensions.}
    \item{\code{cellsize}}{The size of each cell.
Either a single value or one value for each dimension.}
  }}
}
\description{
Add the index to the table
}
