% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index.R
\name{create_index}
\alias{create_index}
\title{Add the index to the table}
\usage{
create_index(grtsdb = getOption("grtsdb", "grts.sqlite"), level)
}
\arguments{
\item{grtsdb}{the grtsdb object}

\item{level}{the required level}
}
\description{
Add the index to the table
}
