% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index.R
\name{create_index}
\alias{create_index}
\title{Add the index to the table}
\usage{
create_index(level, grtsdb = connect_db())
}
\arguments{
\item{level}{the required level}

\item{grtsdb}{the grtsdb object}
}
\description{
Add the index to the table
}
