% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample.R
\name{extract_sample}
\alias{extract_sample}
\title{extract the grts sample}
\usage{
extract_sample(grtsdb = getOption("grtsdb", "grts.sqlite"), samplesize,
  bbox, cellsize)
}
\arguments{
\item{grtsdb}{the grtsdb object}

\item{samplesize}{the required sample size}

\item{bbox}{two-column matrix; the first column has the minimum, the second the maximum values; rows represent the spatial dimensions}

\item{cellsize}{the size of each cell. Either a single value or one value for each dimension.}
}
\description{
extract the grts sample
}
