% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_index.R
\name{has_index}
\alias{has_index}
\title{has a table the required index}
\usage{
has_index(grtsdb, level)
}
\arguments{
\item{grtsdb}{the grtsdb object}

\item{level}{the required level}
}
\description{
has a table the required index
}
