% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grtsdb.R
\name{is_grtsdb}
\alias{is_grtsdb}
\title{Check is an object is a grtsdb}
\usage{
is_grtsdb(grtsdb = getOption("grtsdb", "grts.sqlite"))
}
\arguments{
\item{grtsdb}{the grtsdb object}
}
\description{
Check is an object is a grtsdb
}
